Default initializing all struct/class members
This fixes a crash in debug mode.
The problem was specifically 'm_index' in 'ScriptRenderContext' never being initialized to anything.
In debug mode, it defaults to 0xcccccc which will cause a crash when this variable is used.
I added it to all members because it's good practice to never leave anything uninitialized.